updated pandas numpy sweetviz seaborn shap library versions (#187)
* Fixed a bug in MLflow

* Updated mlfow, pandas and numpy libraries

* Updated requirements.txt to take into account older python version

* Updated pandas version in setup.py to permit 3.8 and 3.9 installation

* Corrected a version to allow numba to be installed

* Updated sweetviz and seaborn

* Corrected a bug in the keras fit (due to update in pandas)

* Same thing on the y_train

* Updated shap library

+++ gabarit/template_nlp/nlp_project/package_name/models_training/models_tensorflow/model_keras.py
@@ -206,8 +206,8 @@ class ModelKeras(ModelClass):
             if isinstance(y_valid, np.ndarray) and len(y_valid.shape) == 2 and y_valid.shape[1] == 1:
                 y_valid = np.ravel(y_valid)
             # Transformation dummies
-            y_train_dummies = pd.get_dummies(y_train)
-            y_valid_dummies = pd.get_dummies(y_valid) if y_valid is not None else None
+            y_train_dummies = pd.get_dummies(y_train).astype(int)
+            y_valid_dummies = pd.get_dummies(y_valid).astype(int) if y_valid is not None else None
             # Important : get_dummies reorder the columns in alphabetical order
             # Thus, there is no problem if we fit again on a new dataframe with shuffled data
             list_classes = list(y_train_dummies.columns)
